+++ test/unit/state/get-drag-impact/combine/should-not-combine-with-home-draggable.spec.js
@@ -7,7 +7,6 @@ import type {
 import { horizontal, vertical } from '../../../../../src/state/axis';
 import getDragImpact from '../../../../../src/state/get-drag-impact';
 import getLiftEffect from '../../../../../src/state/get-lift-effect';
-import { forward } from '../../../../../src/state/user-direction/user-direction-preset';
 import { enableCombining, getPreset } from '../../../../util/dimension';
 
 [vertical, horizontal].forEach((axis: Axis) => {
@@ -34,7 +33,6 @@ import { enableCombining, getPreset } from '../../../../util/dimension';
         droppables: withCombineEnabled,
         previousImpact: homeImpact,
         viewport: preset.viewport,
-        userDirection: forward,
         afterCritical,
       });
       expect(impact).toEqual(homeImpact);
