+++ test/unit/state/get-drag-impact/util/get-center-for-edge.js
